Version 0.4
- Refacto visuliastion (le jeu sera de côté)
- Arbres (produisent des fruits, on peut les secouer pour faire tomber
ceux qui sont murs)
- Reconsolidation des moodles (+famine et deshydratation)
- relief du terrain
- refonte de l'interpolation de la vue cliente
- regroupement des constantes de configuration
- les coordonnées dans la simulations sont basées sur des entites
#onesteptodeterminism
- quelques leger ajustements
- la créature peut se relever pour aller manger ou boire (au lieu de
perdre bêtement de la santé dans son sommeil)
- le froid, le chaud et le mouillé interagissent correctement

+++ src/net/feerie/creatura/shared/commons/Dimension.java
@@ -7,8 +7,8 @@ package net.feerie.creatura.shared.commons;
  */
 public class Dimension
 {
-	public double l;
-	public double h;
+	public int l;
+	public int h;
 	
 	public Dimension()
 	{
@@ -19,7 +19,7 @@ public class Dimension
 	 * @param l la largeur
 	 * @param h la hauteur
 	 */
-	public Dimension(double l, double h)
+	public Dimension(int l, int h)
 	{
 		this.l = l;
 		this.h = h;
